~ more precise timestamp for request_id

+++ lib/lhs/railtie.rb
@@ -8,7 +8,7 @@ module LHS
         private
 
         def prepare_lhs_request_cycle_cache
-          LHS::Record::RequestCycleCache::RequestCycleThreadRegistry.request_id = [Time.now.to_i, request.object_id].join('#')
+          LHS::Record::RequestCycleCache::RequestCycleThreadRegistry.request_id = [Time.now.to_f, request.object_id].join('#')
         end
       end
     end
